# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Orenda API
  description: |
    Turkey-focused "Travel Operating System" — Spring Boot backend.
    POI discovery/search, weather, routing, trip planning, user preferences,
    profile onboarding, and data-import triggers.
  version: 0.0.1-SNAPSHOT
  contact:
    name: Orenda
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - {}
paths:
...
  /api/pois/{id}/reviews:
    get:
      tags: [Reviews]
      summary: Retrieve traveler reviews for a POI
      operationId: listPoiReviews
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Reviews, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PoiReviewResponse'
    post:
      tags: [Reviews]
      summary: Submit a new review/rating for a POI
      operationId: submitPoiReview
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PoiReviewRequest'
      responses:
        '201':
          description: Review created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoiReviewResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: POI not found
...
  /api/trips:
    get:
      tags: [Saved Trips]
      summary: List the authenticated user's saved trips
      operationId: listSavedTrips
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Saved trips (no day detail)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavedTripSummaryResponse'
        '401':
          description: Authentication required
    post:
      tags: [Saved Trips]
      summary: Save a freshly generated plan
      operationId: saveTrip
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveTripRequest'
      responses:
        '201':
          description: Trip saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedTripDetailResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Authentication required
  /api/trips/{id}:
    get:
      tags: [Saved Trips]
      summary: Retrieve a saved trip with full day itineraries and stops
      operationId: getSavedTrip
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Full trip detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedTripDetailResponse'
        '401':
          description: Authentication required
        '404':
          description: Trip not found
    put:
      tags: [Saved Trips]
      summary: Update stops, reorder stops, or adjust notes
      operationId: updateSavedTrip
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTripRequest'
      responses:
        '200':
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedTripDetailResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Trip not found
    patch:
      tags: [Saved Trips]
      summary: Update stops, reorder stops, or adjust notes (alias of PUT)
      operationId: patchSavedTrip
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTripRequest'
      responses:
        '200':
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedTripDetailResponse'
        '404':
          description: Trip not found
    delete:
      tags: [Saved Trips]
      summary: Delete or archive a saved trip
      operationId: deleteSavedTrip
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: archive
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Archive (soft delete) instead of hard delete
      responses:
        '204':
          description: Deleted or archived
        '401':
          description: Authentication required
  /api/trips/{id}/export:
    get:
      tags: [Saved Trips]
      summary: Export trip itinerary to PDF or iCal (.ics)
      operationId: exportSavedTrip
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [ics, pdf]
            default: ics
      responses:
        '200':
          description: File download
          content:
            text/calendar:
              schema:
                type: string
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          description: Unknown export format
        '404':
          description: Trip not found
  /api/trips/recalculate:
    post:
      tags: [Saved Trips]
      summary: Dynamically recalculate remaining stops given a real-time event
      description: Supports RAIN (indoor stops first), DELAY (shift times) and ROAD_CLOSURE (drop unreachable stop). Persists the recalculation and returns the updated trip.
      operationId: recalculateTrip
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecalculateTripRequest'
      responses:
        '200':
          description: Updated trip after recalculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedTripDetailResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Trip or stop not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT returned by POST /api/auth/login or the Google callback.
  schemas:
...
    ApiError:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string

    # ── POI ───────────────────────────────────────────────────────────────
    PoiResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        osmId:
          type: integer
          format: int64
        osmType:
          type: string
          nullable: true
        nameTr:
          type: string
          nullable: true
        nameEn:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        subcategory:
          type: string
          nullable: true
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
        completenessScore:
          type: integer
          format: int32
        distanceKm:
          type: number
          format: double
          nullable: true
          description: Present only in /nearby results
        attributes:
          type: object
          additionalProperties: true
          nullable: true
...
    TripPlanResponse:
      type: object
      properties:
        tripDays:
          type: integer
          format: int32
        summary:
          type: string
        suggestions:
          type: array
          items:
            $ref: '#/components/schemas/ScoredPoi'
        dayPlan:
          type: array
          items:
            $ref: '#/components/schemas/DayPlan'
        notes:
          type: array
          items:
            type: string
        weatherSummary:
          type: string
        preferenceInsight:
          type: string
          nullable: true
        narrative:
          type: string
          nullable: true
    ScoredPoi:
      type: object
      properties:
        poi:
          $ref: '#/components/schemas/PoiResponse'
        score:
          type: number
          format: double
          description: 0–100, higher is better
        factors:
          type: object
          additionalProperties:
            type: number
            format: double
        reasons:
          type: array
          items:
            type: string
        travelMinutes:
          type: integer
          format: int32
          nullable: true
          description: From previous stop / base to this POI
        visitMinutes:
          type: integer
          format: int32
          nullable: true
        startTime:
          type: string
          nullable: true
          example: "10:00"
        endTime:
          type: string
          nullable: true
          example: "11:30"
        openAtScheduledTime:
          type: boolean
          nullable: true
    DayPlan:
      type: object
      properties:
        day:
          type: integer
          format: int32
        date:
          type: string
          format: date
        weather:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/ScoredPoi'
        notes:
          type: array
          items:
            type: string
        narrative:
          type: string
          nullable: true
        lunch:
          $ref: '#/components/schemas/LunchSlot'
    LunchSlot:
      type: object
      properties:
        prompt:
          type: string
        needsDietInfo:
          type: boolean
        nearbyRestaurants:
          type: array
          items:
            $ref: '#/components/schemas/ScoredPoi'
        returnToHotel:
          $ref: '#/components/schemas/ReturnToHotel'
        note:
          type: string
          nullable: true
    ReturnToHotel:
      type: object
      properties:
        travelMinutes:
          type: integer
          format: int32
        distanceKm:
          type: number
          format: double
...
    PoiReviewRequest:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
        title:
          type: string
          maxLength: 200
        comment:
          type: string
          maxLength: 4000
    PoiReviewResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        poiId:
          type: string
          format: uuid
        travelerName:
          type: string
          nullable: true
        rating:
          type: integer
          format: int32
        title:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time

    # ── Saved trips ────────────────────────────────────────────────────────
    SaveTripRequest:
      type: object
      required: [plan]
      properties:
        name:
          type: string
          maxLength: 200
        destination:
          type: string
          maxLength: 200
        plan:
          $ref: '#/components/schemas/TripPlanResponse'
    SavedTripSummaryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        destination:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        tripDays:
          type: integer
          format: int32
        summary:
          type: string
          nullable: true
        weatherSummary:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
    SavedTripDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        destination:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        tripDays:
          type: integer
          format: int32
        summary:
          type: string
          nullable: true
        weatherSummary:
          type: string
          nullable: true
        narrative:
          type: string
          nullable: true
        preferenceInsight:
          type: string
          nullable: true
        notes:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        days:
          type: array
          items:
            $ref: '#/components/schemas/SavedTripDayResponse'
    SavedTripDayResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        day:
          type: integer
          format: int32
        date:
          type: string
          format: date
          nullable: true
        weather:
          type: string
          nullable: true
        narrative:
          type: string
          nullable: true
        notes:
          type: array
          items:
            type: string
        stops:
          type: array
          items:
            $ref: '#/components/schemas/SavedTripStopResponse'
    SavedTripStopResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        poiId:
          type: string
          format: uuid
          nullable: true
        nameTr:
          type: string
          nullable: true
        nameEn:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        subcategory:
          type: string
          nullable: true
        lat:
          type: number
          format: double
          nullable: true
        lon:
          type: number
          format: double
          nullable: true
        score:
          type: number
          format: double
          nullable: true
        travelMinutes:
          type: integer
          format: int32
          nullable: true
        visitMinutes:
          type: integer
          format: int32
          nullable: true
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        openAtScheduledTime:
          type: boolean
          nullable: true
        reasons:
          type: array
          items:
            type: string
        factors:
          type: object
          additionalProperties:
            type: number
            format: double
    UpdateTripRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        destination:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        notes:
          type: array
          items:
            type: string
        days:
          type: array
          items:
            $ref: '#/components/schemas/UpdateTripDay'
    UpdateTripDay:
      type: object
      properties:
        day:
          type: integer
          format: int32
        date:
          type: string
          format: date
          nullable: true
        weather:
          type: string
          nullable: true
        narrative:
          type: string
          nullable: true
        notes:
          type: array
          items:
            type: string
        stops:
          type: array
          items:
            $ref: '#/components/schemas/UpdateTripStop'
    UpdateTripStop:
      type: object
      properties:
        id:
          type: string
          format: uuid
          nullable: true
          description: Existing stop id (optional; the plan is replaced wholesale)
        poiId:
          type: string
          format: uuid
          nullable: true
        nameTr:
          type: string
          nullable: true
        nameEn:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        subcategory:
          type: string
          nullable: true
        lat:
          type: number
          format: double
          nullable: true
        lon:
          type: number
          format: double
          nullable: true
        score:
          type: number
          format: double
          nullable: true
        travelMinutes:
          type: integer
          format: int32
          nullable: true
        visitMinutes:
          type: integer
          format: int32
          nullable: true
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        openAtScheduledTime:
          type: boolean
          nullable: true
        reasons:
          type: array
          items:
            type: string
        factors:
          type: object
          additionalProperties:
            type: number
            format: double
    RecalculateTripRequest:
      type: object
      required: [tripId, event]
      properties:
        tripId:
          type: string
          format: uuid
        event:
          type: string
          enum: [RAIN, DELAY, ROAD_CLOSURE]
        eventDate:
          type: string
          format: date
          nullable: true
          description: Day from which the recalculation applies (null = whole trip)
        delayMinutes:
          type: integer
          format: int32
          minimum: 1
          nullable: true
          description: Required for DELAY
        affectedStopId:
          type: string
          format: uuid
          nullable: true
          description: Required for ROAD_CLOSURE
        note:
          type: string
          nullable: true